refactor(core): move platform providers to services
