+++ scripts/core-boundaries/rules/crate-rules.mjs
@@ -460,6 +460,7 @@ export const dependencyProfileRules = [
       'dunce',
       'futures',
       'git2',
+      'local-ip-address',
       'notify',
       'rand',
       'reqwest',
@@ -470,6 +471,7 @@ export const dependencyProfileRules = [
       'tokio-util',
       'tokio-tungstenite',
       'uuid',
+      'which',
       'bitfun-relay-server',
     ],
   },
