+++ scripts/core-boundaries/rules/feature-rules.mjs
@@ -27,7 +27,6 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'image', ownerFeatures: ['service-integrations'] },
       { depName: 'include_dir', ownerFeatures: ['product-full'] },
       { depName: 'indexmap', ownerFeatures: ['product-full'] },
-      { depName: 'local-ip-address', ownerFeatures: ['service-integrations'] },
       { depName: 'md5', ownerFeatures: ['product-full', 'service-integrations'] },
       { depName: 'rand', ownerFeatures: ['service-integrations'] },
       { depName: 'reqwest', ownerFeatures: ['ai-adapter-runtime', 'service-integrations'] },
@@ -60,20 +59,21 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'bitfun-runtime-ports', ownerFeatures: ['remote-connect'] },
       {
         depName: 'bitfun-services-core',
-        ownerFeatures: ['git', 'mcp', 'miniapp-runtime', 'workspace-search'],
+        ownerFeatures: ['git', 'mcp', 'miniapp-runtime', 'remote-connect', 'workspace-search'],
       },
       { depName: 'chrono', ownerFeatures: ['git', 'remote-connect', 'remote-ssh-concrete'] },
-      { depName: 'dirs', ownerFeatures: ['miniapp-runtime', 'remote-ssh-concrete'] },
+      { depName: 'dirs', ownerFeatures: ['miniapp-runtime', 'remote-connect', 'remote-ssh-concrete'] },
       { depName: 'dunce', ownerFeatures: ['remote-ssh', 'workspace-search'] },
       { depName: 'futures', ownerFeatures: ['mcp', 'remote-connect'] },
       { depName: 'git2', ownerFeatures: ['git'] },
       { depName: 'hostname', ownerFeatures: ['remote-connect'] },
       { depName: 'image', ownerFeatures: ['remote-connect'] },
+      { depName: 'local-ip-address', ownerFeatures: ['remote-connect'] },
       { depName: 'mac_address', ownerFeatures: ['remote-connect'] },
       { depName: 'notify', ownerFeatures: ['file-watch'] },
       { depName: 'qrcode', ownerFeatures: ['remote-connect'] },
       { depName: 'rand', ownerFeatures: ['mcp', 'remote-connect', 'remote-ssh-concrete'] },
-      { depName: 'reqwest', ownerFeatures: ['mcp', 'miniapp-runtime'] },
+      { depName: 'reqwest', ownerFeatures: ['announcement', 'mcp', 'miniapp-runtime', 'remote-connect'] },
       { depName: 'rmcp', ownerFeatures: ['mcp'] },
       { depName: 'russh', ownerFeatures: ['remote-ssh-concrete'] },
       { depName: 'russh-keys', ownerFeatures: ['remote-ssh-concrete'] },
@@ -91,7 +91,7 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'tokio-util', ownerFeatures: ['remote-ssh'] },
       { depName: 'urlencoding', ownerFeatures: ['remote-connect'] },
       { depName: 'uuid', ownerFeatures: ['miniapp-runtime', 'remote-connect', 'remote-ssh-concrete'] },
-      { depName: 'which', ownerFeatures: ['miniapp-runtime', 'workspace-search'] },
+      { depName: 'which', ownerFeatures: ['miniapp-runtime', 'remote-connect', 'workspace-search'] },
       { depName: 'x25519-dalek', ownerFeatures: ['remote-connect'] },
     ],
   },
