+++ src/crates/services/services-integrations/src/announcement/mod.rs
@@ -1,8 +1,10 @@
 //! Announcement service data contracts.
 
+mod remote;
 mod state_store;
 mod types;
 
+pub use remote::{AnnouncementRemoteFetchRequest, RemoteAnnouncementFetcher};
 pub use state_store::{
     AnnouncementStateStore, AnnouncementStateStoreError, AnnouncementStateStoreResult,
 };
